+++ build_scripts/generate_api_docs.py
@@ -4,11 +4,11 @@ from pathlib import Path
 import mkdocs_gen_files
 
 nav = mkdocs_gen_files.Nav()
-
-for path in sorted(Path("src").rglob("*.py")):
-    module_path = path.relative_to("src").with_suffix("")
-    doc_path = path.relative_to("src").with_suffix(".md")
-    full_doc_path = Path("code-reference") / doc_path
+root = Path("src")  # / Path("pydvl")
+for path in sorted(root.rglob("*.py")):
+    module_path = path.relative_to(root).with_suffix("")
+    doc_path = path.relative_to(root).with_suffix(".md")
+    full_doc_path = Path("api") / doc_path
     parts = tuple(module_path.parts)
 
     if parts[-1] == "__init__":
@@ -26,5 +26,5 @@ for path in sorted(Path("src").rglob("*.py")):
 
     mkdocs_gen_files.set_edit_path(full_doc_path, path)
 
-with mkdocs_gen_files.open("code-reference/SUMMARY.md", "w") as nav_file:
-    nav_file.writelines(nav.build_literate_nav())
+# with mkdocs_gen_files.open("api/SUMMARY.md", "w") as nav_file:
+#     nav_file.writelines(nav.build_literate_nav())