+++ src/pydvl/utils/parallel/futures/__init__.py
@@ -2,10 +2,13 @@ from concurrent.futures import Executor
 from contextlib import contextmanager
 from typing import Generator, Optional
 
-from joblib.externals.loky import get_reusable_executor
-
 from pydvl.utils.config import ParallelConfig
-from pydvl.utils.parallel.futures.ray import RayExecutor
+from pydvl.utils.parallel.backend import BaseParallelBackend
+
+try:
+    from pydvl.utils.parallel.futures.ray import RayExecutor
+except ImportError:
+    pass
 
 __all__ = ["init_executor"]
 
@@ -14,9 +17,9 @@ __all__ = ["init_executor"]
 def init_executor(
     max_workers: Optional[int] = None,
     config: ParallelConfig = ParallelConfig(),
-    **kwargs: dict,
+    **kwargs,
 ) -> Generator[Executor, None, None]:
-    """Initializes a futures executor based on the passed parallel configuration object.
+    """Initializes a futures executor for the given parallel configuration.
 
     Args:
         max_workers: Maximum number of concurrent tasks.
@@ -24,35 +27,27 @@ def init_executor(
             with cluster address, number of cpus, etc.
         kwargs: Other optional parameter that will be passed to the executor.
 
-    Examples:
 
-    >>> from pydvl.utils.parallel.futures import init_executor
-    >>> from pydvl.utils.config import ParallelConfig
-    >>> config = ParallelConfig(backend="ray")
-    >>> with init_executor(max_workers=3, config=config) as executor:
-    ...     pass
-
-    >>> from pydvl.utils.parallel.futures import init_executor
-    >>> with init_executor() as executor:
-    ...     future = executor.submit(lambda x: x + 1, 1)
-    ...     result = future.result()
-    ...
-    >>> print(result)
-    2
-
-    >>> from pydvl.utils.parallel.futures import init_executor
-    >>> with init_executor() as executor:
-    ...     results = list(executor.map(lambda x: x + 1, range(5)))
-    ...
-    >>> print(results)
-    [1, 2, 3, 4, 5]
-
+    ??? Examples
+        ``` python
+        from pydvl.utils.parallel.futures import init_executor
+        from pydvl.utils.config import ParallelConfig
+        config = ParallelConfig(backend="ray")
+        with init_executor(max_workers=1, config=config) as executor:
+            future = executor.submit(lambda x: x + 1, 1)
+            result = future.result()
+        assert result == 2
+        ```
+        ``` python
+        from pydvl.utils.parallel.futures import init_executor
+        with init_executor() as executor:
+            results = list(executor.map(lambda x: x + 1, range(5)))
+        assert results == [1, 2, 3, 4, 5]
+        ```
     """
-    if config.backend == "ray":
-        with RayExecutor(max_workers, config=config, **kwargs) as executor:
-            yield executor
-    elif config.backend == "joblib":
-        with get_reusable_executor(max_workers=max_workers, **kwargs) as executor:
-            yield executor
-    else:
-        raise NotImplementedError(f"Unexpected parallel type {config.backend}")
+    try:
+        cls = BaseParallelBackend.BACKENDS[config.backend]
+        with cls.executor(max_workers=max_workers, config=config, **kwargs) as e:
+            yield e
+    except KeyError:
+        raise NotImplementedError(f"Unexpected parallel backend {config.backend}")