+++ src/pydvl/value/result.py
@@ -28,7 +28,7 @@ the same way.
 Indexing and slicing of results is supported and
 [ValueItem][pydvl.value.result.ValueItem] objects are returned. These objects
 can be compared with the usual operators, which take only the
-[ValueItem.value][pydvl.value.result.ValueItem.value] into account.
+[ValueItem.value][pydvl.value.result.ValueItem] into account.
 
 # Creating result objects
 
@@ -232,7 +232,7 @@ class ValuationResult(
         algorithm: str = "",
         status: Status = Status.Pending,
         sort: bool = False,
-        **extra_values: dict,
+        **extra_values,
     ):
         if variances is not None and len(variances) != len(values):
             raise ValueError("Lengths of values and variances do not match")
@@ -293,7 +293,7 @@ class ValuationResult(
         Args:
             reverse: Whether to sort in descending order by value.
             key: The key to sort by. Defaults to
-                [ValueItem.value][pydvl.value.result.ValueItem.value].
+                [ValueItem.value][pydvl.value.result.ValueItem].
         """
         keymap = {
             "index": "_indices",
@@ -678,7 +678,7 @@ class ValuationResult(
 
     @classmethod
     def from_random(
-        cls, size: int, total: Optional[float] = None, **kwargs: dict
+        cls, size: int, total: Optional[float] = None, **kwargs
     ) -> "ValuationResult":
         """Creates a [ValuationResult][pydvl.value.result.ValuationResult] object and fills it with an array
         of random values from a uniform distribution in [-1,1]. The values can
@@ -693,7 +693,7 @@ class ValuationResult(
 
         Returns:
             A valuation result with its status set to
-                [Status.Converged][Status.Converged] by default.
+            [Status.Converged][pydvl.utils.status.Status] by default.
 
         :raises ValueError: If `size` is less than 1.