+++ src/pydvl/value/shapley/montecarlo.py
@@ -33,83 +33,89 @@ or using an early stopping strategy to reduce computation
    the algorithms mentioned above, including Group Testing, can work to valuate
    groups of samples as units.
 """
+from __future__ import annotations
+
 import logging
 import math
 import operator
+from concurrent.futures import FIRST_COMPLETED, Future, wait
 from functools import reduce
 from itertools import cycle, takewhile
 from typing import Sequence
 
 import numpy as np
+from deprecate import deprecated
 from numpy.typing import NDArray
 from tqdm import tqdm
 
+from pydvl.utils import effective_n_jobs, init_executor, init_parallel_backend
 from pydvl.utils.config import ParallelConfig
 from pydvl.utils.numeric import random_powerset
-from pydvl.utils.parallel import MapReduceJob
+from pydvl.utils.parallel import CancellationPolicy, MapReduceJob
 from pydvl.utils.utility import Utility
 from pydvl.value.result import ValuationResult
 from pydvl.value.shapley.truncated import NoTruncation, TruncationPolicy
-from pydvl.value.stopping import StoppingCriterion
+from pydvl.value.stopping import MaxChecks, StoppingCriterion
 
 logger = logging.getLogger(__name__)
 
 __all__ = ["permutation_montecarlo_shapley", "combinatorial_montecarlo_shapley"]
 
 
-def _permutation_montecarlo_shapley(
-    u: Utility,
-    *,
-    done: StoppingCriterion,
-    truncation: TruncationPolicy,
-    algorithm_name: str = "permutation_montecarlo_shapley",
-    progress: bool = False,
-    job_id: int = 1,
+def _permutation_montecarlo_one_step(
+    u: Utility, truncation: TruncationPolicy, algorithm_name: str
 ) -> ValuationResult:
     """Helper function for [permutation_montecarlo_shapley()][pydvl.value.shapley.montecarlo.permutation_montecarlo_shapley].
 
-    Computes marginal utilities of each training sample in
-    [Utility.data][pydvl.utils.utility.Utility.data] by iterating through
-    randomly sampled permutations.
-
+    Computes marginal utilities of each training sample in a randomly sampled
+    permutation.
     Args:
         u: Utility object with model, data, and scoring function
-        done: Check on the results which decides when to stop
         truncation: A callable which decides whether to interrupt
             processing a permutation and set all subsequent marginals to zero.
         algorithm_name: For the results object. Used internally by different
             variants of Shapley using this subroutine
-        progress: Whether to display progress bars for each job.
-        job_id: id to use for reporting progress (e.g. to place progres bars)
 
     Returns:
         An object with the results
     """
+
     result = ValuationResult.zeros(
         algorithm=algorithm_name, indices=u.data.indices, data_names=u.data.data_names
     )
 
-    pbar = tqdm(disable=not progress, position=job_id, total=100, unit="%")
-    while not done(result):
-        pbar.n = 100 * done.completion()
-        pbar.refresh()
-        prev_score = 0.0
-        permutation = np.random.permutation(u.data.indices)
-        permutation_done = False
-        truncation.reset()
-        for i, idx in enumerate(permutation):
-            if permutation_done:
-                score = prev_score
-            else:
-                score = u(permutation[: i + 1])
-            marginal = score - prev_score
-            result.update(idx, marginal)
-            prev_score = score
-            if not permutation_done and truncation(i, score):
-                permutation_done = True
+    prev_score = 0.0
+    permutation = np.random.permutation(u.data.indices)
+    permutation_done = False
+    truncation.reset()
+    for i, idx in enumerate(permutation):
+        if permutation_done:
+            score = prev_score
+        else:
+            score = u(permutation[: i + 1])
+        marginal = score - prev_score
+        result.update(idx, marginal)
+        prev_score = score
+        if not permutation_done and truncation(i, score):
+            permutation_done = True
+    nans = np.isnan(result.values).sum()
+    if nans > 0:
+        logger.warning(
+            f"{nans} NaN values in current permutation, ignoring. "
+            "Consider setting a default value for the Scorer"
+        )
+        result = ValuationResult.empty(algorithm=algorithm_name)
     return result
 
 
+@deprecated(
+    target=True,
+    deprecated_in="0.7.0",
+    remove_in="0.8.0",
+    args_mapping=dict(
+        coordinator_update_period=None, worker_update_period=None, progress=None
+    ),
+)
 def permutation_montecarlo_shapley(
     u: Utility,
     done: StoppingCriterion,
@@ -119,16 +125,39 @@ def permutation_montecarlo_shapley(
     config: ParallelConfig = ParallelConfig(),
     progress: bool = False,
 ) -> ValuationResult:
-    r"""Computes an approximate Shapley value by sampling independent index
-    permutations to approximate the sum:
+    r"""Computes an approximate Shapley value by sampling independent
+    permutations of the index set, approximating the sum:
 
     $$
     v_u(x_i) = \frac{1}{n!} \sum_{\sigma \in \Pi(n)}
     \tilde{w}( | \sigma_{:i} | )[u(\sigma_{:i} \cup \{i\}) − u(\sigma_{:i})],
     $$
 
-    where $\sigma_{:i}$ denotes the set of indices in permutation sigma before the
-    position where $i$ appears (see [Data valuation][computing-data-values] for details).
+    where $\sigma_{:i}$ denotes the set of indices in permutation sigma before
+    the position where $i$ appears (see [[data-valuation]] for details).
+
+    This implements the method described in [@ghorbani_data_2019]
+    with a double stopping criterion.
+
+    .. todo::
+       Think of how to add Robin-Gelman or some other more principled stopping
+       criterion.
+
+    Instead of naively implementing the expectation, we sequentially add points
+    to coalitions from a permutation and incrementally compute marginal utilities.
+    We stop computing marginals for a given permutation based on a
+    [TruncationPolicy][pydvl.value.shapley.truncated.TruncationPolicy].
+    [@ghorbani_data_2019] mention two policies: one that stops after a certain
+    fraction of marginals are computed, implemented in
+    [FixedTruncation][pydvl.value.shapley.truncated.FixedTruncation],
+    and one that stops if the last computed utility ("score") is close to the
+    total utility using the standard deviation of the utility as a measure of
+    proximity, implemented in
+    [BootstrapTruncation][pydvl.value.shapley.truncated.BootstrapTruncation].
+
+    We keep sampling permutations and updating all shapley values
+    until the [StoppingCriterion][pydvl.value.stopping.StoppingCriterion] returns
+    `True`.
 
     Args:
         u: Utility object with model, data, and scoring function.
@@ -139,26 +168,47 @@ def permutation_montecarlo_shapley(
         n_jobs: number of jobs across which to distribute the computation.
         config: Object configuring parallel computation, with cluster address,
             number of cpus, etc.
-        progress: Whether to display progress bars for each job.
+        progress: Whether to display a progress bar.
 
     Returns:
         Object with the data values.
     """
+    algorithm = "permutation_montecarlo_shapley"
 
-    map_reduce_job: MapReduceJob[Utility, ValuationResult] = MapReduceJob(
-        u,
-        map_func=_permutation_montecarlo_shapley,
-        reduce_func=lambda results: reduce(operator.add, results),
-        map_kwargs=dict(
-            algorithm_name="permutation_montecarlo_shapley",
-            done=done,
-            truncation=truncation,
-            progress=progress,
-        ),
-        config=config,
-        n_jobs=n_jobs,
-    )
-    return map_reduce_job()
+    parallel_backend = init_parallel_backend(config)
+    u = parallel_backend.put(u)
+    max_workers = effective_n_jobs(n_jobs, config)
+    n_submitted_jobs = 2 * max_workers  # number of jobs in the executor's queue
+
+    result = ValuationResult.zeros(algorithm=algorithm)
+
+    pbar = tqdm(disable=not progress, total=100, unit="%")
+
+    with init_executor(
+        max_workers=max_workers, config=config, cancel_futures=CancellationPolicy.ALL
+    ) as executor:
+        pending: set[Future] = set()
+        while True:
+            pbar.n = 100 * done.completion()
+            pbar.refresh()
+
+            completed, pending = wait(
+                pending, timeout=config.wait_timeout, return_when=FIRST_COMPLETED
+            )
+
+            for future in completed:
+                result += future.result()
+                # we could check outside the loop, but that means more
+                # submissions if the stopping criterion is unstable
+                if done(result):
+                    return result
+
+            # Ensure that we always have n_submitted_jobs in the queue or running
+            for _ in range(n_submitted_jobs - len(pending)):
+                future = executor.submit(
+                    _permutation_montecarlo_one_step, u, truncation, algorithm
+                )
+                pending.add(future)
 
 
 def _combinatorial_montecarlo_shapley(
@@ -227,7 +277,7 @@ def combinatorial_montecarlo_shapley(
     \binom{n-1}{ | S | }^{-1} [u(S \cup \{i\}) − u(S)]$$
 
     This consists of randomly sampling subsets of the power set of the training
-    indices in [data][pydvl.utils.utility.Utility.data], and computing their
+    indices in [u.data][pydvl.utils.utility.Utility], and computing their
     marginal utilities. See [Data valuation][computing-data-values] for details.
 
     Note that because sampling is done with replacement, the approximation is