+++ src/pydvl/value/shapley/owen.py
@@ -42,8 +42,8 @@ def _owen_sampling_shapley(
     Args:
         indices: Indices to compute the value for
         u: Utility object with model, data, and scoring function
-        method: Either [Full][pydvl.value.shapley.owen.OwenAlgorithm.Full] for
-            q ∈ [0, 1] or [Halved][pydvl.value.shapley.owen.OwenAlgorithm.Halved]
+        method: Either [OwenAlgorithm.Full][pydvl.value.shapley.owen.OwenAlgorithm]
+            for q ∈ [0, 1] or [OwenAlgorithm.Halved][pydvl.value.shapley.owen.OwenAlgorithm]
             for q ∈ [0, 0.5] and correlated samples
         n_samples: Number of subsets to sample to estimate the integrand
         max_q: number of subdivisions for the integration over $q$
@@ -136,8 +136,9 @@ def owen_sampling_shapley(
         max_q: Number of subdivisions for q ∈ [0,1] (the element sampling
             probability) used to approximate the outer integral.
         method: Selects the algorithm to use, see the description. Either
-            [Full][pydvl.value.shapley.owen.OwenAlgorithm.Full] for $q \in [0,1]$
-            or [Halved][pydvl.value.shapley.owen.OwenAlgorithm.Halved] for
+            [OwenAlgorithm.Full][pydvl.value.shapley.owen.OwenAlgorithm] for
+            $q \in [0,1]$ or
+            [OwenAlgorithm.Halved][pydvl.value.shapley.owen.OwenAlgorithm] for
             $q \in [0,0.5]$ and correlated samples
         n_jobs: Number of parallel jobs to use. Each worker receives a chunk
             of the total of `max_q` values for q.