+++ tests/value/test_sampler.py
@@ -6,8 +6,8 @@ import pytest
 from pydvl.utils import powerset
 from pydvl.value.sampler import (
     AntitheticSampler,
-    DeterministicCombinatorialSampler,
     DeterministicPermutationSampler,
+    DeterministicUniformSampler,
     PermutationSampler,
     RandomHierarchicalSampler,
     UniformSampler,
@@ -17,7 +17,7 @@ from pydvl.value.sampler import (
 @pytest.mark.parametrize(
     "sampler_class",
     [
-        DeterministicCombinatorialSampler,
+        DeterministicUniformSampler,
         UniformSampler,
         PermutationSampler,
         AntitheticSampler,
@@ -38,7 +38,7 @@ def test_proper(sampler_class, indices):
 @pytest.mark.parametrize(
     "sampler_class",
     [
-        DeterministicCombinatorialSampler,
+        DeterministicUniformSampler,
         UniformSampler,
         #    PermutationSampler,
         AntitheticSampler,